+++ src/app/test/e2e/restaurant/conftest.py
@@ -1,0 +1,1 @@
+import pytest
